........
-updated to asio 1.4.8
........
-add 64-bit VS2010 project settings for reTurn
........
-merge work from b-counterpath-recon-20130424 with some changes
 - upped receive buffer size from 2048 to 4096
 - added setOnBeforeSocketClosedFp to AsyncSocketBase - Provides an opportunity
   for the app to clean up, e.g., QoS-related data or resources just before 
   the socket is closed
 - OS level Udp socket receive buffer size set to 66560
 - TurnAsyncSocketBase - use dispatch instead of post for send API for
   increased efficiency
 - TurnAsyncSocket - new client side APIs
   - setLocalPassword for checking integrity of incoming STUN messages
   - connectivityCheck for ICE connectivity checks 
   - setOnBeforeSocketClosedFp for QOS cleanup (windows)
   - send API now split into send(To)Framed and send(To)Unframed
   - Queue of guards modified to use a weak functor template instead
   - onBindFailure and onBindSuccess now return the Tuple that failed or succeeded
 - DataBuffer improvements
 - StunMessage - added ice attributes

........
-track allocations per connection - reduces map sizes and lookups when TCP/TLS 
 client connections are used
........
-fixup stunTestVectors.cxx now that ice parameters are parsed
........
-fix up TestRTPLoad and add VS2010 project for it
........
-fix handling of tryAgain flag for websockets transport - buffer pointers and
 sizes weren't being adjusted - which led to heap corruption
-the weProcessData method needs a good review and optimization pass - it 
 contains a number of data copies that should be able to be optimized out
........
-changed TurnAyncSocket to use asio dispatch instead of post for better efficiency
 when call fns from the ioService thread.
-fix initialization issue in TestRtpLoad
-fix project settings in TestRTPLoad
-fix for requestTimeout - handle case where close is called in handlers and it
 is invalidated
........
recon: update for using installed sipX media library
........
recon: test case includes
........
recon: tests: do not run testUA automatically
........
recon: look for dynamic codec modules in default location if none are statically linked
........
recon: test: build testUA, use AM_CPPFLAGS
........
-changes to VS2010 project settings to be more consistent
........
-changes to VS2010 project settings to be more consistent

........
recon: test return values, log more progress
........
reflow: ensure proper initialization of policy structure
........
recon: revise earlier patch, just log failures creating SRTP session (FIXME, should drop participant)
........
recon: logging when mMediaStream not ready/SRTP setup fails
........
recon: testUA: add commented logging setup for sipX
........
recon: more warnings about media setup failure
........
reflow: emit warning when socket is not available
........
recon: testUA: add commented G.722 support (depends on sipXmediaLib fix)
........
reflow: only display warning when necessary
........
resip/stack: SDP: support (S)AVPF
........
recon: add SAVPF fudge support
........
recon: MOHParkServer: ensure headers included in correct order, sipX is very sensitive to this
........
recon: MOHParkServer: specify dependency locations for autotools build
........
resip/stack: ConnectionBase: WS manage mBufferPos
........
resip/stack: ConnectionBase: WS eliminate copy
........
resip/stack: ConnectionBase: WS eliminate c_str() usage that may result in copy, fix string termination
........
-reverted Daniels commit to change mBufferPos - it has reintroduced heap corruption
 - you need to change reduce bytesRead by the number of bytes processed already
   when setting tryAgain to true
 - the remaining buffer is copied back to position 0 (see for statement 
   above changed code), so mBufferPos should be 0
   - ideally we would entirely remove this buffer copy
 - in else clause mBufferPos shouldn't be 0 - there may be a partial frame left
   over in the buffer
 - this section of code needs an in-depth review

........
recon: testUA: elaborate on G.722 situation
........
resip/stack: class for efficiently parsing WebSocket streams (work in progress)
........
resip/stack: add missing headers to includes list
........
resip/stack: add debug msg about fragment SIP messages in WebSockets
........
resip/stack: WebSockets: use standalone class for parsing, code optimization
........
-allow AllowEvents to be an advertised capability
........
removed some compiler warnings and added new WsFrameExtractor to VS build files
........
-added VS2010 project files for MOHParkServer
-replaced deprecated bzero calls with memsets since bzero doesn't exist in Visual Studio compiler

........
rutil: Data: import capacity of pre-allocated buffers
........
resip/stack: WebSocket: allocate extra space in buffers for null bytes.  Ensure Data ShareEnum is not changed to Take by doing resize manually
........
resip/recon/MOHParkServer: ensure Windows build artifacts are distributed
........
resip/recon/MOHParkServer: ensure Windows build artifacts are distributed (correction)
........
resip/recon: ensure playback_prompt.h is distributed
........
resip/recon: ensure record_prompt.h is distributed
........
resip/stack: TcpConnection: adjust logging levels
........


git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-counterpath-recon-20130424@10203 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ reTurn/AsyncTcpSocketBase.cxx
@@ -29,7 +29,7 @@ AsyncTcpSocketBase::~AsyncTcpSocketBase()
 unsigned int 
 AsyncTcpSocketBase::getSocketDescriptor() 
 { 
-   return mSocket.native(); 
+   return (unsigned int)mSocket.native(); 
 }
 
 asio::error_code 
@@ -205,7 +205,7 @@ AsyncTcpSocketBase::transportClose()
 {
    if (mOnBeforeSocketCloseFp)
    {
-      mOnBeforeSocketCloseFp(mSocket.native());
+      mOnBeforeSocketCloseFp((unsigned int)mSocket.native());
    }
 
    asio::error_code ec;
