+++ reTurn/client/TurnAsyncTlsSocket.hxx
@@ -25,7 +25,7 @@ public:
                                const asio::ip::address& address = UnspecifiedIpAddress, 
                                unsigned short port = 0);
 
-   virtual unsigned int getSocketDescriptor() { return mSocket.lowest_layer().native(); }
+   virtual unsigned int getSocketDescriptor() { return (unsigned int)mSocket.lowest_layer().native(); }
 
 protected:
